requirements.txt
Flask==2.0.2
speedtest-cli==2.1.3
